+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## <small>2.3.42 (2025-01-18)</small>
+
+* fix ts ([7649cb9](https://github.com/justdlabs/cli/commit/7649cb9))
+
+## <small>2.3.41 (2025-01-18)</small>
+
+* fix: fix stubs ([4918fb6](https://github.com/justdlabs/cli/commit/4918fb6))
+* wip ([7b87f6f](https://github.com/justdlabs/cli/commit/7b87f6f))
+
+## <small>2.3.40 (2025-01-18)</small>
+
+* fix: providers and theme provider ([96fe206](https://github.com/justdlabs/cli/commit/96fe206))
+
+## <small>2.3.39 (2025-01-18)</small>
+
+* bump ([f2150c0](https://github.com/justdlabs/cli/commit/f2150c0))
+* fix: javascript ([71da248](https://github.com/justdlabs/cli/commit/71da248))
+
 ## <small>2.3.36 (2025-01-18)</small>
 
 * fix: remove separator from sidebar ([267ae1b](https://github.com/justdlabs/cli/commit/267ae1b))